basic_q_and_a/src/chunker.py
from langchain_text_splitters import RecursiveCharacterTextSplitter


def split_documents(documents, chunk_size: int = 600, chunk_overlap: int = 100):
    """Split documents into smaller chunks for better RAG performance"""
    
    text_splitter = RecursiveCharacterTextSplitter(
        chunk_size=chunk_size,
        chunk_overlap=chunk_overlap,
        length_function=len,
        separators=["\n\n","\n"," ",""]
    )
    
    split_docs = text_splitter.split_documents(documents)
    print(f"Split {len(documents)} documents into {len(split_docs)} chunks")
    
    if split_docs:
        print(f"\n[DEBUG] Example Chunk: ")
        print(f"[DEBUG] Content: {split_docs[0].page_content[:200]}...")
        print(f"[DEBUG] Metadata: {split_docs[0].metadata}")
        
        
    return split_docs
    
    