basic_q_and_a/pyproject.toml
[project]
name = "basic-q-and-a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.2",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.0",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "pymupdf>=1.26.6",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "typesense>=1.1.1",
    "uvicorn>=0.38.0",
]
